Extract shared FFC and organizer components
