+++ jitx-py/shared-components/src/shared_components/__init__.py
@@ -1,11 +1,15 @@
-from shared_components.ffc import HDGC60PinFfc
+from shared_components.ffc import HDGC60PinFfc, RetroBus60FfcConnector
 from shared_components.saleae import SaleaeProbeHeader2x4, SignalGroundHeader2x4
+from shared_components.sharp_organizer import SharpOrganizerBus, SharpOrganizerHost
 from shared_components.testpads import GndTestpads, SignalTestPad
 
 __all__ = [
     "GndTestpads",
     "HDGC60PinFfc",
+    "RetroBus60FfcConnector",
     "SaleaeProbeHeader2x4",
+    "SharpOrganizerBus",
+    "SharpOrganizerHost",
     "SignalGroundHeader2x4",
     "SignalTestPad",
 ]
